+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # CHANGELOG
 
+#### master
+
+Incompatible changes:
+
+- CHANGED: Domain.ExpiresOn has been replaced by Domain.ExpiresAt. (dnsimple/dnsimple-go#98)
+
 #### Release 0.63.0
+
 - NEW: Added types and parsing for account membership events. (dnsimple/dnsimple-go#97)